youtube-posting/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
Pillow>=10.0.0
anthropic>=0.86.0
gtts>=2.3.0
